msm: vidc: raise interrupt for encode batching usecase
HFR recording usecases, batch mode is enabled by default.
Currently no interrupt is raised to firmware for ETB write
on command queue. So firmware was not reading the queue,
though there are lot of pending messages in the command
queue, So complete encoder pipeline is getting stalled
after encoding couple of frames.

To avoid above problem, raising interrupt to firmware
every time(Only for last ETB in the batch).

Change-Id: If584a8175bfa54402087d1888dd032d08e1f3c15
Signed-off-by: Govindaraj Rajagopal <grajagop@codeaurora.org>