msm: vidc: Don't add extra 4K alignment for linear color fmts
For linear color fmts extra 4K alignment was being added to the
total calculated size which leads to mismatch between gralloc
alloc size and filled length size.

Instead of adding the alignment directly, MSM_MEDIA_ALIGN should
be used to align the total buffer size if the latter is not
already 4K aligned.

Change-Id: I0a3c65a65780e94837617244ede28336a411250e