video: driver: Fix OOB access while handling fence property
Fence property payload is copied into a fixed length instance
array. When the payload exceeds the size of array, there is a
possible case of out of bound access.
The fix ensures that the memcpy is avoided when the payload
exceed the size of memory allocated to accomodate the payload.

Change-Id: I38b8c0f0ee92011f6209adfb83c7b9c0fd869e4a
Signed-off-by: Vikash Garodia <quic_vgarodia@quicinc.com>
(cherry picked from commit 3ccf4735eaac1532a824db0a6bb61168e2793d78)